Extract structs into models.go file
